fix: ConvertFacet `RemoveDeposit` int128 -> int96

+++ protocol/contracts/beanstalk/silo/ConvertFacet.sol
@@ -37,7 +37,7 @@ contract ConvertFacet is ReentrancyGuard {
     event RemoveDeposit(
         address indexed account,
         address indexed token,
-        int128 stem,
+        int96 stem,
         uint256 amount,
         uint256 bdv
     );
